# backend/openapi.yaml
openapi: 3.0.3
info:
  title: MeroSadak API
  description: |
    Nepal's intelligent road monitoring and safety platform.
    Provides real-time road conditions, weather alerts, route planning, and incident reporting.
  version: 1.0.0
  contact:
    name: MeroSadak Team
    email: support@merosadak.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:4000/api/v1
    description: Development server
  - url: https://merosadak.onrender.com/api/v1
    description: Production server
...
paths:
...
  /highways/{code}/incidents:
    get:
      tags: [Highways]
      summary: Get incidents for specific highway
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Highway incidents
...
  /incidents:
    get:
      tags: [Incidents]
      summary: Get recent user-reported incidents
      responses:
        '200':
          description: List of incidents
    post:
      tags: [Incidents]
      summary: Submit incident report
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [type, description, lat, lng]
              properties:
                type:
                  type: string
                  enum: [blockage, accident, traffic, construction, weather, other]
                description:
                  type: string
                  minLength: 10
                  maxLength: 1000
                lat:
                  type: number
                  minimum: -90
                  maximum: 90
                lng:
                  type: number
                  minimum: -180
                  maximum: 180
                severity:
                  type: string
                  enum: [low, medium, high, critical]
      responses:
        '201':
          description: Incident reported
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...